RD-833 Split app.json into smaller files, updated RPM spec (#1158)
