+++ backend/config.js
@@ -5,7 +5,10 @@ const flatten = require('flat');
 
 const Utils = require('./utils');
 
+const root = require('../conf/config.json');
 const app = require('../conf/app.json');
+const logging = require('../conf/logging.json');
+const dbOptions = require('../conf/db.options.json');
 const manager = require('../conf/manager.json');
 let userConfig = require('../conf/userConfig.json');
 
@@ -32,7 +35,7 @@ try {
 module.exports = {
     get(mode) {
         const config = {
-            app: _.merge(app, userConfig),
+            app: _.merge(app, root, logging, { db: { options: dbOptions } }, userConfig),
             manager,
             mode
         };
