+++ conf/config.json
@@ -1,0 +1,37 @@
+{
+    "maintenancePollingInterval": 10000,
+    "singleManager": true,
+
+    "proxy": {
+        "timeouts": {
+            "get": 10000,
+            "post": 30000,
+            "put": 30000,
+            "delete": 10000,
+            "blueprintUpload": 70000
+        }
+    },
+    "github": {
+        "username": "secret(github-username)",
+        "password": "secret(github-password)"
+    },
+    "source": {
+        "browseSourcesDir": "/cloudifyBrowseSources",
+        "lookupYamlsDir": "/cloudifyLookupYamls"
+    },
+    "widgets": {
+        "ignoreFolders": ["common"],
+        "tempDir": "cloudifyWidgets",
+        "requiredFiles": ["widget.js", "widget.png"],
+        "backendFilename": "backend.js"
+    },
+    "ssl": {
+        "ca": "/etc/cloudify/ssl/cloudify_internal_ca_cert.pem"
+    },
+    "saml":{
+        "enabled": false,
+        "certPath": "",
+        "ssoUrl": "",
+        "portalUrl": ""
+    }
+}
