+++ conf/logging.json
@@ -1,0 +1,7 @@
+{
+    "logsFile": "/var/log/cloudify/stage/server-output.log",
+    "errorsFile": "/var/log/cloudify/stage/server-error.log",
+    "logLevelConf": "/etc/cloudify/logging.conf",
+    "serviceName": "cloudify-stage",
+    "logLevel": "warn"
+}
